requirements.txt
imageio
numpy
scipy
pyyaml
easydict